Reorganized request handling
Lookup now functions similarly to assembler - a data structure with the command name and a function pointer contains all implemented debug adapter requests and their handlers. The handle_request function now simply looks up the right function and invokes it, nothing else required.

This simplifies the request handler by a great deal, and should make working on the project less intimidating.

This comes with the side effect of making it less clear where *exactly* the child name-emu process is spawned and terminated. This is a tradeoff I'm willing to make, as the request handler is the most important part of this binary and should be as simple as possible.